Move service code into its own package since it does not have to be tied to the server technologies (e.g., echo).  This reduces dependencies for other services that re-use this code.
