+++ service/service_test.go
@@ -1,4 +1,4 @@
-package server
+package service
 
 import (
 	"encoding/json"
